+++ spec/mysql-client_spec.rb
@@ -9,12 +9,12 @@ describe 'openstack-ops-database::mysql-client' do
     let(:node) { runner.node }
     let(:chef_run) { runner.converge(described_recipe) }
 
-    it 'includes mysql client recipes' do
-      expect(chef_run).to include_recipe 'mysql::client'
+    it 'has default mysql client resource' do
+      expect(chef_run).to create_mysql_client 'default'
     end
 
-    it 'includes mysql-chef_gem recipes' do
-      expect(chef_run).to include_recipe 'mysql-chef_gem::default'
+    it 'has default mysql chef gem resource' do
+      expect(chef_run).to install_mysql2_chef_gem 'default'
     end
 
     it 'installs mysql packages' do