+++ CHANGELOG.md
@@ -2,6 +2,9 @@ openstack-ops-database Cookbook CHANGELG
 ===================================
 This file is used to list changes made in each version of the openstack-ops-database cookbook.
 
+## 10.1.0
+* Updates for mysql cookbook ver 6.x
+
 ## 10.0.0
 * Upgrading to Juno
 * Upgrading berkshelf from 2.0.18 to 3.1.5